refactor: enhance filter resolution in dashboard components
- Introduced a new utility function, resolveFilterParams, to streamline the resolution of Handlebars templates in pgrest parameter values based on active filters.
- Updated multiple components, including PlannerProvider, useDashletData, and usePgrestResolvedFields, to utilize the new function for improved clarity and maintainability in filter management.
- This change enhances the overall filtering logic and ensures that only valid parameters are passed to pgrest queries.
