+++ apps/app/src/features/dashboard/dashlets/common/use-pgrest-resolved-fields.ts
@@ -1,11 +1,11 @@
 import { useMemo } from "react";
 import type { PgrestParam, PgrestHttpMethod } from "./pgrest-types";
+import { EMPTY_PGREST_PARAMS } from "./pgrest-types";
 import { usePgrestRows } from "./use-pgrest-rows";
 import { usePlannerData } from "./use-planner-data";
 import { resolveHandlebarsField } from "./use-handlebars-templates";
+import { resolveFilterParams } from "./resolve-filter-params";
 import { useDashboardFilters } from "../../context/dashboard-filters-context";
-
-const EMPTY_PARAMS: PgrestParam[] = [];
 
 interface PgrestResolvedFieldsConfig {
   dataMode: "static" | "pgrest" | "planner";
@@ -40,26 +40,12 @@ export function usePgrestResolvedFields({
   const { activeFilters } = useDashboardFilters();
 
   // Resolve {{filter.*}} templates in pgrest param values before fetching
-  const resolvedParams = useMemo(() => {
-    if (pgrestParams.length === 0) return EMPTY_PARAMS;
-    const filterContext = { filter: activeFilters };
-    const resolved = pgrestParams.map((p) => {
-      if (p.value?.includes("{{filter.")) {
-        const resolvedValue = resolveHandlebarsField(p.value, filterContext);
-        return { ...p, value: resolvedValue, _fromTemplate: true };
-      }
-      return { ...p, _fromTemplate: false };
-    });
-    return resolved
-      .filter((p) => {
-        if (!p._fromTemplate) return true;
-        // Drop template params that resolved to empty or operator-only (e.g. "eq.", "like.")
-        return p.value !== "" && !p.value.endsWith(".");
-      })
-      .map(({ _fromTemplate, ...p }) => p);
-  }, [pgrestParams, activeFilters]);
+  const resolvedParams = useMemo(
+    () => resolveFilterParams(pgrestParams, activeFilters),
+    [pgrestParams, activeFilters],
+  );
 
-  const stableParams = resolvedParams.length > 0 ? resolvedParams : EMPTY_PARAMS;
+  const stableParams = resolvedParams.length > 0 ? resolvedParams : EMPTY_PGREST_PARAMS;
 
   const { rows: pgrestRowsResult, loading: pgrestLoading, fetchError: pgrestError } = usePgrestRows(
     dataMode === "pgrest" ? "pgrest" : "static",
